More refactor + more tests
